get rid of half and add resize embeddings to trianing script

+++ examples/pytorch/text_to_speech/run_vits_finetuning.py
@@ -404,8 +404,6 @@ def discriminator_loss(disc_real_outputs, disc_generated_outputs):
     real_losses = 0
     generated_losses = 0
     for disc_real, disc_generated in zip(disc_real_outputs, disc_generated_outputs):
-        disc_real = disc_real.float()
-        disc_generated = disc_generated.float()
         real_loss = torch.mean((1 - disc_real) ** 2)
         generated_loss = torch.mean(disc_generated**2)
         loss += real_loss + generated_loss
@@ -419,8 +417,7 @@ def feature_loss(feature_maps_real, feature_maps_generated):
     loss = 0
     for feature_map_real, feature_map_generated in zip(feature_maps_real, feature_maps_generated):
         for real, generated in zip(feature_map_real, feature_map_generated):
-            real = real.float().detach()
-            generated = generated.float()
+            real = real.detach()
             loss += torch.mean(torch.abs(real - generated))
 
     return loss * 2
@@ -430,7 +427,7 @@ def generator_loss(disc_outputs):
     total_loss = 0
     gen_losses = []
     for disc_output in disc_outputs:
-        disc_output = disc_output.float()
+        disc_output = disc_output
         loss = torch.mean((1 - disc_output) ** 2)
         gen_losses.append(loss)
         total_loss += loss
@@ -443,11 +440,6 @@ def kl_loss(prior_latents, posterior_log_variance, prior_means, prior_log_varian
     z_p, logs_q: [b, h, t_t]
     prior_means, prior_log_variance: [b, h, t_t]
     """
-    prior_latents = prior_latents.float()
-    posterior_log_variance = posterior_log_variance.float()
-    prior_means = prior_means.float()
-    prior_log_variance = prior_log_variance.float()
-    labels_mask = labels_mask.float()
 
     kl = prior_log_variance - posterior_log_variance - 0.5
     kl += 0.5 * ((prior_latents - prior_means) ** 2) * torch.exp(-2.0 * prior_log_variance)
@@ -1135,7 +1127,7 @@ def main():
                 _, fmaps_target = discriminator(target_waveform)
                 discriminator_candidate, fmaps_candidate = discriminator(waveform)
 
-                loss_duration = torch.sum(log_duration.float())
+                loss_duration = torch.sum(log_duration)
                 loss_mel = torch.nn.functional.l1_loss(mel_scaled_target, mel_scaled_generation)
                 loss_kl = kl_loss(
                     prior_latents, posterior_log_variances, prior_means, prior_log_variances, labels_padding_mask
